requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
requests>=2.31.0
resend>=2.0.0
playwright>=1.40.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
responses>=0.23.0
freezegun>=1.2.0
